% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{date_r}
\alias{date_r}
\alias{date_acc}
\title{Date Received}
\usage{
date_r(pubmed_article_element)

date_acc(pubmed_article_element)
}
\arguments{
\item{pubmed_article_element}{An XML node containing a PubMed article element.}
}
\value{
A character string representing the received date or "N/A" if not found.

A character string representing the accepted date or "N/A" if not found.
}
\description{
This function extracts the received date from a PubMed article element.

This function extracts the accepted date from a PubMed article element.
}
