% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_funcs.R
\name{fetch_pubmed_content}
\alias{fetch_pubmed_content}
\title{Fetch PubMed Content}
\usage{
fetch_pubmed_content(
  db = "pubmed",
  ids = NULL,
  retmode = "abstract",
  rettype = "xml"
)
}
\arguments{
\item{db}{A character string specifying the database to search in. Default is "pubmed".}

\item{ids}{A character vector specifying the PubMed article IDs for which to retrieve content.}

\item{retmode}{A character string specifying the desired format of the retrieved content. Default is "abstract".}

\item{rettype}{A character string specifying the type of content to retrieve. Default is "xml".}
}
\value{
An XML document containing the content of the specified PubMed articles.
}
\description{
This internal function retrieves content from PubMed articles based on a list of PubMed IDs.
}
\keyword{internal}
