% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlink.R
\name{process_pubmed_outlinks}
\alias{process_pubmed_outlinks}
\title{Process PubMed Outlinks}
\usage{
process_pubmed_outlinks(xml_doc)
}
\arguments{
\item{xml_doc}{An XML document object representing the PubMed search results.}
}
\value{
A data frame containing columns "PMID" and "Outlink", where "PMID" represents the PubMed ID and "Outlink" represents the outlink URL.
}
\description{
This function processes the outlinks from a PubMed XML document and returns a data frame containing PMIDs and their corresponding outlink URLs.
}
\keyword{internal}
